backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
litellm
duckduckgo-search
tavily-python
httpx
pydantic
